+++ migratie/migr-test-persoon-naarlo3/src/main/java/nl/bzk/migratiebrp/test/brpnaarlo3/adapter/entity/PersoonNaamgebruikHistorieConverter.java
@@ -8,14 +8,16 @@ package nl.bzk.migratiebrp.test.brpnaarlo3.adapter.entity;
 
 import java.sql.Timestamp;
 import java.util.Set;
-import nl.bzk.migratiebrp.synchronisatie.dal.domein.brp.kern.entity.AdellijkeTitel;
-import nl.bzk.migratiebrp.synchronisatie.dal.domein.brp.kern.entity.BRPActie;
-import nl.bzk.migratiebrp.synchronisatie.dal.domein.brp.kern.entity.Naamgebruik;
-import nl.bzk.migratiebrp.synchronisatie.dal.domein.brp.kern.entity.Persoon;
-import nl.bzk.migratiebrp.synchronisatie.dal.domein.brp.kern.entity.PersoonNaamgebruikHistorie;
-import nl.bzk.migratiebrp.synchronisatie.dal.domein.brp.kern.entity.Predicaat;
+
+import nl.bzk.algemeenbrp.dal.domein.brp.entity.BRPActie;
+import nl.bzk.algemeenbrp.dal.domein.brp.entity.Persoon;
+import nl.bzk.algemeenbrp.dal.domein.brp.entity.PersoonNaamgebruikHistorie;
+import nl.bzk.algemeenbrp.dal.domein.brp.enums.AdellijkeTitel;
+import nl.bzk.algemeenbrp.dal.domein.brp.enums.Naamgebruik;
+import nl.bzk.algemeenbrp.dal.domein.brp.enums.Predicaat;
 import nl.bzk.migratiebrp.test.brpnaarlo3.adapter.ConverterContext;
 import nl.bzk.migratiebrp.test.brpnaarlo3.adapter.OnbekendeHeaderException;
+
 import org.springframework.stereotype.Component;
 
 /**
@@ -96,19 +98,19 @@ public final class PersoonNaamgebruikHistorieConverter extends AbstractEntityHis
                 persoon = context.getPersoon(Integer.parseInt(value));
                 break;
             case HEADER_WIJZE_GEBRUIK_GESLACHTSNAAM:
-                naamgebruik = Naamgebruik.parseId(Short.valueOf(value));
+                naamgebruik = Naamgebruik.parseId(Integer.valueOf(value));
                 break;
             case HEADER_INDICATIE_NAAMGEBRUIK_AFGELEID:
                 indicatieNaamgebruikAfgeleid = Boolean.valueOf(value);
                 break;
             case HEADER_PREDICAAT_NAAMGEBRUIK:
-                predikaat = Predicaat.parseId(Short.valueOf(value));
+                predikaat = Predicaat.parseId(Integer.valueOf(value));
                 break;
             case HEADER_VOORNAMEN_NAAMGEBRUIK:
                 voornamenNaamgebruik = value;
                 break;
             case HEADER_ADELLIJKETITEL_NAAMGEBRUIK:
-                adellijkeTitel = AdellijkeTitel.parseId(Short.valueOf(value));
+                adellijkeTitel = AdellijkeTitel.parseId(Integer.valueOf(value));
                 break;
             case HEADER_VOORVOEGSEL_NAAMGEBRUIK:
                 voorvoegselNaamgebruik = value;
